Address comments in #713
