+++ api/query/plan.go
@@ -1,0 +1,8 @@
+// Copyright 2018 The WPT Dashboard Project. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+package query
+
+// TODO: Establish query execution plans over internal data schema.
+type plan interface{}
